fix: address PR feedback — remove double goBack, use utils.dispatch, extract MVVM
- Remove redundant navigation.goBack() in ServiceOutage (onSuccess already dismisses)
- Use utils.dispatch for stale banner cleanup (already in useCallback deps via utils)
- Extract useServiceOutageViewModel hook following project MVVM pattern
